fix: correct filtering for MAP types
