+++ src/components/Item/VisualizationItem/Visualization/Visualization.js
@@ -3,7 +3,7 @@ import i18n from '@dhis2/d2-i18n'
 import { Button, Cover, IconInfo24, colors } from '@dhis2/ui'
 import uniqueId from 'lodash/uniqueId.js'
 import PropTypes from 'prop-types'
-import React, { useMemo, useCallback } from 'react'
+import React, { useMemo } from 'react'
 import { useSelector } from 'react-redux'
 import {
     VISUALIZATION,
@@ -59,9 +59,14 @@ const Visualization = ({
         [visualization, activeType, originalType]
     )
 
-    const getCbFilteredVisualization = useCallback(
-        () => getFilteredVisualization(visualizationConfig, itemFilters),
-        [visualizationConfig, itemFilters]
+    const filteredVisualization = useMemo(
+        () =>
+            getFilteredVisualization(
+                visualizationConfig,
+                itemFilters,
+                originalType
+            ),
+        [visualizationConfig, itemFilters, originalType]
     )
 
     const filterVersion = useMemo(() => uniqueId(), [])
@@ -78,14 +83,14 @@ const Visualization = ({
             itemType: item.type,
         }),
         [
+            originalType,
             activeType,
+            style,
+            filterVersion,
             dashboardMode,
             dashboardId,
-            filterVersion,
             item.id,
             item.type,
-            originalType,
-            style,
         ]
     )
 
@@ -99,7 +104,7 @@ const Visualization = ({
         case VISUALIZATION: {
             return (
                 <IframePlugin
-                    visualization={getCbFilteredVisualization()}
+                    visualization={filteredVisualization}
                     {...iFramePluginProps}
                 />
             )
@@ -135,8 +140,7 @@ const Visualization = ({
         case MAP: {
             return (
                 <MapPlugin
-                    visualization={visualizationConfig}
-                    itemFilters={itemFilters}
+                    visualization={filteredVisualization}
                     {...iFramePluginProps}
                 />
             )
